backend/requirements.txt
flask
pandas
numpy
python-dotenv
reportlab
openpyxl
matplotlib
pytz
gunicorn
sqlalchemy
psycopg2-binary
scikit-learn